Final push

+++ springboot-project/src/main/java/com/pfg/controllers/HomeController.java
@@ -22,6 +22,7 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
@@ -42,44 +43,55 @@ import org.springframework.web.context.request.ServletRequestAttributes;
 import org.springframework.web.multipart.MultipartFile;
 
 @Controller
+@RequestMapping("/home")
 public class HomeController {
 
+    //SERVICIO DE USUARIO
     @Autowired
     private IUserService service;
 
+    //SERVICIO DE INTERESES
     @Autowired
     private IInterestService intService;
 
+    //SERVICIO DE DATOS DE USUARIO
     @Autowired
     private IUserDataService userDataService;
 
+    //CLASE SISTEMA DE GUARDADO DE ARCHIVOS
     @Autowired
     private UploadFileService uploadService;
 
+    //SERVICIO DE EVENTOS
     @Autowired
     private IEventService eventService;
 
+    //CLASE SISTEMA DE SESIONES
     private SessionUtils SU = new SessionUtils();
 
+    //ENCRIPTADOR PARA CONTRASEÑAS
     private BCryptPasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
 
-    @GetMapping({ "/home/contact" })
+    //PAGINA DE CONTACTO
+    @GetMapping({ "/contact" })
     public String contactPage(Model model) {
         return "contact";
     }
 
+    //OBTENER SESION ACTUAL
     public HttpSession getSession(){
         ServletRequestAttributes attr = (ServletRequestAttributes) RequestContextHolder.currentRequestAttributes();
         return attr.getRequest().getSession(false);
     }
 
+    //OBTENER USUARIO DE LA SESION
     public User getSessionUser() {
-        ServletRequestAttributes attr = (ServletRequestAttributes) RequestContextHolder.currentRequestAttributes();
-        HttpSession session = attr.getRequest().getSession(false);
+        HttpSession session = getSession();
         User sessionUser = (User) session.getAttribute("user");
         return sessionUser;
     }
 
+    //INTERESES EN COMUN ENTRE DOS USUARIOS
     private String getCommonInterestsFraction(User user1, User user2) {
         List<Interest> user1Interests = intService.listByIndexes(userDataService.getInterestList(user1));
         List<Interest> user2Interests = intService.listByIndexes(userDataService.getInterestList(user2));
@@ -90,8 +102,15 @@ public class HomeController {
         return commonInterestsCount + "/" + user1Interests.size() + " intereses coincidentes";
     }
 
+    //PAGINA DE DESCUBRIMIENTO DE PERSONAS
     @GetMapping({ "/meet/{id}" })
     public String meetPage(Model model, @PathVariable Long id) {
+        //COMPROBACION DE SESION
+        boolean sessionN = SU.checkSession(getSession());
+        if (!sessionN) {
+            return "redirect:/";
+        }
+        //---------------------
         User user = service.readUserId(id);
         List<Interest> currentUserInterests = intService.listByIndexes(userDataService.getInterestList(user));
         List<User> matchedUsers = service.findUsersByInterests(currentUserInterests);
@@ -119,6 +138,7 @@ public class HomeController {
         return "meet_page";
     }
 
+    //INTERESES COMUNES ENTRE USUARIOS
     private int countCommonInterests(User user, List<Interest> currentUserInterests) {
         List<Interest> userInterests = intService.listByIndexes(userDataService.getInterestList(user));
         int commonInterests = 0;
@@ -145,14 +165,14 @@ public class HomeController {
         return friends;
     }
 
-    // Pagina de inicio de sesion
-    @GetMapping({ "/users/login" })
+    //PAGINA DE INICIO DE SESION
+    @GetMapping({ "/login" })
     public String redirectLogIn(Model model) {
         return "try_session";
     }
 
-    // Cierre de sesion
-    @GetMapping({ "/users/logout" })
+    //CIERRE DE SESION
+    @GetMapping({ "/logout" })
     public String logout(Model model) {
         ServletRequestAttributes attr = (ServletRequestAttributes) RequestContextHolder.currentRequestAttributes();
         HttpSession session = attr.getRequest().getSession(false);
@@ -160,8 +180,8 @@ public class HomeController {
         return "redirect:/";
     }
 
-    // Filtrado de usuarios
-    @PostMapping("/users/filter")
+    //FILTRADO DE USUARIOS
+    @PostMapping("/filter")
     public ResponseEntity<String> viewUsersByFilter(@Valid @RequestBody Map<String, String> payload) {
         String interestValue = payload.get("value");
         Long id = Long.parseLong(interestValue); // ID RECOGIDO
@@ -192,9 +212,15 @@ public class HomeController {
         }
     }
 
-    // PAGINA DE USUARIO POR INTERES
-    @GetMapping("/users/view/{id}")
+    //PAGINA DE USUARIO POR INTERES
+    @GetMapping("/view/{id}")
     public String getMethodName(Model model, @PathVariable Long id) {
+        //COMPROBACION DE SESION
+        boolean sessionN = SU.checkSession(getSession());
+        if (!sessionN) {
+            return "redirect:/";
+        }
+        //---------------------
         User user = service.readUserId(id);
         model.addAttribute("user", user);
         model.addAttribute("birth", user.getFormattedBirthdate());
@@ -202,17 +228,17 @@ public class HomeController {
         return "view_user";
     }
 
-    // REPORTAR USUARIO
-    @GetMapping("/users/report/{id}")
+    //REPORTAR USUARIO
+    @GetMapping("/report/{id}")
     public String getMethodName(@PathVariable Long id) {
         User user = service.readUserId(id);
         UserData userD = user.getUserData();
         userD.setReport_number(userD.getReport_number() + 1);
         userDataService.saveUserPreferences(userD);
-        return "redirect:/meet/" + getSessionUser().getId();
+        return "redirect:/home/meet/" + getSessionUser().getId();
     }
 
-    // Pagina de usuario
+    //PAGINA DE USUARIO
     @GetMapping({ "/userpage/{id}" })
     public String userPage(Model model, @PathVariable Long id) {
         //COMPROBACION DE SESION
@@ -247,15 +273,8 @@ public class HomeController {
         return "user_page";
     }
 
-    // Listado de usuarios
-    @GetMapping({ "/users" })
-    public String listUsers(Model model) {
-        model.addAttribute("users", service.listAllUsers());
-        return "users";
-    }
-
-    // Creacion de nuevos usuarios
-    @GetMapping("/users/new")
+    //CREACION DE NUEVOS USUARIOS
+    @GetMapping("/new")
     public String showRegistration(Model model) {
         User user = new User();
         int[] avatarList = { 1, 2, 3, 4, 5, 6 };
@@ -265,6 +284,7 @@ public class HomeController {
         return "create_user";
     }
 
+
     @PostMapping("/users")
     public String createUser(@ModelAttribute("user") User user, BindingResult bindingResult, HttpServletRequest request,
             @RequestParam("profileImage") MultipartFile file) {
@@ -273,7 +293,7 @@ public class HomeController {
             userCreated = service.readEmail(user.getEmail());
         }
         if (userCreated != null) {
-            return "redirect:/users/new";
+            return "redirect:/home/new";
         } else {
             uploadService.saveUserImage(file);
             user.setImage_url(file.getOriginalFilename());
@@ -301,47 +321,57 @@ public class HomeController {
                 if (session == null) { // CASO SI ES UN USUARIO NUEVO DESDE EL INDEX
                     session = attr.getRequest().getSession(true);
                     session.setAttribute("user", user);
-                    return "redirect:/userpage/" + user.getId();
-                } else {
-                    User sessionUser = (User) session.getAttribute("user");
-                    if (sessionUser.getRol().isAdministrator()) {
-                        return "redirect:/users";
-                    } else {
-                        return "redirect:/userpage/" + user.getId();
-                    }
+                    
                 }
-
             }
-            return "redirect:/users";
+            return "redirect:/home/userpage/" + user.getId();
         }
     }
 
-    // Borrado de usuarios
-    @GetMapping("/users/{id}")
+    //BORRADO DE USUARIOS
+    @GetMapping("/{id}")
     public String deleteUser(@PathVariable Long id) {
+        //COMPROBACION DE SESION
+        boolean sessionN = SU.checkSession(getSession());
+        if (!sessionN) {
+            return "redirect:/";
+        }
+        //---------------------
         User user = service.readUserId(id);
         ServletRequestAttributes attr = (ServletRequestAttributes) RequestContextHolder.currentRequestAttributes();
         HttpSession session = attr.getRequest().getSession(false);
         User sessionUser = (User) session.getAttribute("user");
         if (sessionUser.getRol().isAdministrator() && sessionUser.getId() != user.getId()) {
             service.deleteUser(id);
-            return "redirect:/userpage/" + sessionUser.getId();
+            return "redirect:/home/userpage/" + sessionUser.getId();
         } else {
             service.deleteUser(id);
             return "redirect:/";
         }
     }
 
-    // Actualizacion de usuarios
-    @GetMapping("/users/edit/{id}")
+    //ACTUALIZACION DE USUARIOS
+    @GetMapping("/edit/{id}")
     public String showRegistration(@PathVariable Long id, Model model) {
+        //COMPROBACION DE SESION
+        boolean sessionN = SU.checkSession(getSession());
+        if (!sessionN) {
+            return "redirect:/";
+        }
+        //---------------------
         User user = service.readUserId(id);
         model.addAttribute("user", user);
         return "edit_user";
     }
 
-    @PostMapping("/users/update")
+    @PostMapping("/update")
     public String updateUser(@ModelAttribute("user") User user, BindingResult bindingResult) {
+        //COMPROBACION DE SESION
+        boolean sessionN = SU.checkSession(getSession());
+        if (!sessionN) {
+            return "redirect:/";
+        }
+        //---------------------
         User existingUser = service.readUserId(user.getId());
         existingUser.setUsername(user.getUsername());
         existingUser.setName(user.getName());
@@ -355,12 +385,12 @@ public class HomeController {
         ServletRequestAttributes attr = (ServletRequestAttributes) RequestContextHolder.currentRequestAttributes();
         HttpSession session = attr.getRequest().getSession(false);
         User sessionUser = (User) session.getAttribute("user");
-        return "redirect:/userpage/" + sessionUser.getId();
+        return "redirect:/home/userpage/" + sessionUser.getId();
 
     }
 
-    // Comprobacion para inicio de sesion
-    @PostMapping("/users/checkuser")
+    //CHECKEO DE SESION DE USUARIO
+    @PostMapping("/checkuser")
     public String checkUser(@ModelAttribute("user") User user, BindingResult bindingResult, Model model) {
         User userC = service.readUserName(user.getUsername());
         if (bindingResult.hasErrors()) {
@@ -370,7 +400,7 @@ public class HomeController {
             ServletRequestAttributes attr = (ServletRequestAttributes) RequestContextHolder.currentRequestAttributes();
             HttpSession session = attr.getRequest().getSession(true);
             session.setAttribute("user", userC);
-            return "redirect:/userpage/" + userC.getId();
+            return "redirect:/home/userpage/" + userC.getId();
 
         } else {
             return "try_session";
